BigDecimal - added support for fractional hours

+++ src/main/java/pl/edu/agh/mwo/reporter/model/report/Report2.java
@@ -1,11 +1,12 @@
 package pl.edu.agh.mwo.reporter.model.report;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
 public class Report2 {
     private final String description;
-    private final Map<String, Integer> hoursPerProject;
+    private final Map<String, BigDecimal> hoursPerProject;
 
     public Report2(String description) {
         this.description = description;
@@ -16,15 +17,15 @@ public class Report2 {
         return description;
     }
 
-    public Map<String, Integer> getHoursPerProject() {
+    public Map<String, BigDecimal> getHoursPerProject() {
         return hoursPerProject;
     }
 
-    public Integer getHoursForProject(String projectName) {
+    public BigDecimal getHoursForProject(String projectName) {
         return this.hoursPerProject.get(projectName);
     }
 
-    public void addProjectWithTotalNumberOfHours(String projectName, Integer totalNumberOfHours) {
+    public void addProjectWithTotalNumberOfHours(String projectName, BigDecimal totalNumberOfHours) {
         this.hoursPerProject.put(projectName, totalNumberOfHours);
     }
 }
